+++ Tests/test_file_png.py
@@ -88,20 +88,13 @@ class TestFilePng(PillowTestCase):
         self.assertEqual(im.format, "PNG")
         self.assertEqual(im.get_format_mimetype(), 'image/png')
 
-        hopper("1").save(test_file)
-        Image.open(test_file)
-
-        hopper("L").save(test_file)
-        Image.open(test_file)
-
-        hopper("P").save(test_file)
-        Image.open(test_file)
-
-        hopper("RGB").save(test_file)
-        Image.open(test_file)
-
-        hopper("I").save(test_file)
-        Image.open(test_file)
+        for mode in ["1", "L", "P", "RGB", "I", "I;16"]:
+            im = hopper(mode)
+            im.save(test_file)
+            reloaded = Image.open(test_file)
+            if mode == "I;16":
+                reloaded = reloaded.convert(mode)
+            self.assert_image_equal(reloaded, im)
 
     def test_invalid_file(self):
         invalid_file = "Tests/images/flower.jpg"
@@ -589,6 +582,40 @@ class TestFilePng(PillowTestCase):
         # Raises an EOFError in load_end
         im = Image.open("Tests/images/hopper_idat_after_image_end.png")
         self.assertEqual(im.text, {'TXT': 'VALUE', 'ZIP': 'VALUE'})
+
+    def test_exif(self):
+        im = Image.open("Tests/images/exif.png")
+        exif = im._getexif()
+        self.assertEqual(exif[274], 1)
+
+    def test_exif_save(self):
+        im = Image.open("Tests/images/exif.png")
+
+        test_file = self.tempfile("temp.png")
+        im.save(test_file)
+
+        reloaded = Image.open(test_file)
+        exif = reloaded._getexif()
+        self.assertEqual(exif[274], 1)
+
+    def test_exif_from_jpg(self):
+        im = Image.open("Tests/images/pil_sample_rgb.jpg")
+
+        test_file = self.tempfile("temp.png")
+        im.save(test_file)
+
+        reloaded = Image.open(test_file)
+        exif = reloaded._getexif()
+        self.assertEqual(exif[305], "Adobe Photoshop CS Macintosh")
+
+    def test_exif_argument(self):
+        im = Image.open(TEST_PNG_FILE)
+
+        test_file = self.tempfile("temp.png")
+        im.save(test_file, exif=b"exifstring")
+
+        reloaded = Image.open(test_file)
+        self.assertEqual(reloaded.info["exif"], b"Exif\x00\x00exifstring")
 
     @unittest.skipUnless(HAVE_WEBP and _webp.HAVE_WEBPANIM,
                          "WebP support not installed with animation")