+++ src/PIL/Image.py
@@ -578,7 +578,12 @@ class Image(object):
         return self
 
     def __exit__(self, *args):
-        self.close()
+        if hasattr(self, "_close__fp"):
+            self._close__fp()
+        if (hasattr(self, 'fp') and hasattr(self, '_exclusive_fp')
+           and self.fp and self._exclusive_fp):
+            self.fp.close()
+        self.fp = None
 
     def close(self):
         """
@@ -610,12 +615,7 @@ class Image(object):
 
     if sys.version_info.major >= 3:
         def __del__(self):
-            if hasattr(self, "_close__fp"):
-                self._close__fp()
-            if (hasattr(self, 'fp') and hasattr(self, '_exclusive_fp')
-               and self.fp and self._exclusive_fp):
-                self.fp.close()
-            self.fp = None
+            self.__exit__()
 
     def _copy(self):
         self.load()
@@ -1049,7 +1049,7 @@ class Image(object):
                 new_im.info['transparency'] = trns
         return new_im
 
-    def quantize(self, colors=256, method=None, kmeans=0, palette=None):
+    def quantize(self, colors=256, method=None, kmeans=0, palette=None, dither=1):
         """
         Convert the image to 'P' mode with the specified number
         of colors.
@@ -1062,6 +1062,10 @@ class Image(object):
         :param kmeans: Integer
         :param palette: Quantize to the palette of given
                         :py:class:`PIL.Image.Image`.
+        :param dither: Dithering method, used when converting from
+           mode "RGB" to "P" or from "RGB" or "L" to "1".
+           Available methods are NONE or FLOYDSTEINBERG (default).
+           Default: 1 (legacy setting)
         :returns: A new image
 
         """
@@ -1089,7 +1093,7 @@ class Image(object):
                 raise ValueError(
                     "only RGB or L mode images can be quantized to a palette"
                     )
-            im = self.im.convert("P", 1, palette.im)
+            im = self.im.convert("P", dither, palette.im)
             return self._new(im)
 
         return self._new(self.im.quantize(colors, method, kmeans))
@@ -2002,7 +2006,7 @@ class Image(object):
         library automatically seeks to frame 0.
 
         Note that in the current version of the library, most sequence
-        formats only allows you to seek to the next frame.
+        formats only allow you to seek to the next frame.
 
         See :py:meth:`~PIL.Image.Image.tell`.
 
@@ -2021,10 +2025,10 @@ class Image(object):
         debugging purposes.
 
         On Unix platforms, this method saves the image to a temporary
-        PPM file, and calls either the **xv** utility or the **display**
+        PPM file, and calls the **display**, **eog** or **xv**
         utility, depending on which one can be found.
 
-        On macOS, this method saves the image to a temporary BMP file, and
+        On macOS, this method saves the image to a temporary PNG file, and
         opens it with the native Preview application.
 
         On Windows, it saves the image to a temporary BMP file, and uses