+++ src/PIL/PngImagePlugin.py
@@ -529,6 +529,11 @@ class PngStream(ChunkStream):
 
         return s
 
+    def chunk_eXIf(self, pos, length):
+        s = ImageFile._safe_read(self.fp, length)
+        self.im_info["exif"] = b"Exif\x00\x00"+s
+        return s
+
     # APNG chunks
     def chunk_acTL(self, pos, length):
         s = ImageFile._safe_read(self.fp, length)
@@ -683,6 +688,12 @@ class PngImageFile(ImageFile.ImageFile):
         self.png.close()
         self.png = None
 
+    def _getexif(self):
+        if "exif" not in self.info:
+            self.load()
+        from .JpegImagePlugin import _getexif
+        return _getexif(self)
+
 
 # --------------------------------------------------------------------
 # PNG writer
@@ -696,6 +707,7 @@ _OUTMODES = {
     "L":    ("L",       b'\x08\x00'),
     "LA":   ("LA",      b'\x08\x04'),
     "I":    ("I;16B",   b'\x10\x00'),
+    "I;16": ("I;16B",   b'\x10\x00'),
     "P;1":  ("P;1",     b'\x01\x03'),
     "P;2":  ("P;2",     b'\x02\x03'),
     "P;4":  ("P;4",     b'\x04\x03'),
@@ -860,6 +872,12 @@ def _save(im, fp, filename, chunk=putchunk):
             if cid in chunks:
                 chunks.remove(cid)
                 chunk(fp, cid, data)
+
+    exif = im.encoderinfo.get("exif", im.info.get("exif"))
+    if exif:
+        if exif.startswith(b"Exif\x00\x00"):
+            exif = exif[6:]
+        chunk(fp, b"eXIf", exif)
 
     ImageFile._save(im, _idat(fp, chunk),
                     [("zip", (0, 0)+im.size, 0, rawmode)])